Move each loader closer to their route component
