+++ src/renderer/views/Settings/SettingsAudio.tsx
@@ -4,7 +4,8 @@ import { useRouteLoaderData } from 'react-router-dom';
 import * as Setting from '../../components/Setting/Setting';
 import AudioOutputSelect from '../../components/AudioOutputSelect/AudioOutputSelect';
 import { usePlayerAPI } from '../../stores/usePlayerStore';
-import { SettingsLoaderResponse } from '../router';
+
+import { SettingsLoaderResponse } from './Settings';
 
 export default function SettingsAudio() {
   const { config } = useRouteLoaderData('settings') as SettingsLoaderResponse;
